🧹 Consolidate duplicate formatDuration in Prow hooks
Extract identical formatDuration(startTime, endTime?) from useProw.ts
and useCachedProw.ts into shared formatProwDuration in lib/formatters.ts.
Replace magic numbers (1000, 60) with time constants from lib/constants/time.ts.

Fixes #10279
